Fix Cucumber/Watir/Cheezy locator for edit_ve
$ bundle exec cucumber features/visual_editor_logged_in.feature
was failing with
  unable to locate element, using {:text=>"Edit", :tag_name=>"a"}
  (Watir::Exception::UnknownObjectException)

because I assume, it's searching for text: 'Edit' and the VE tab text is
(temporarily) Edit<sup>beta</sup>.  The safest way to locate the <a> is
by the parent tab's id, but it's nested differently in different skins,
so it seems I'm forced to use xpath[1].  This works, but the test then
dies because VE puts up a big warning/welcome alert. Baby steps.

[1] The test framework that adopts jQuery selectors will do well.

Change-Id: If65af0a355e394b25f940eb16e279a90001ecd6a