Monitor fatal errors on beta labs
We discussed doing this at the quarterly review:
https://www.mediawiki.org/wiki/Wikimedia_Release_and_QA_Team/Quarterly_review,_November_2013

This script is intended to run each hour.  In its current form
it could be started as a cron job.  Alternately, we could have
it loop and sleep for seconds_since_last_run while running in
a screen() session.

I tested this by running it on deployment-bastion and sending
messages to myself, and it works nicely.

Change-Id: I0498bb87dd3cb239ea08d521ddf285faea9cffee